erc20ForwarderClient object
